Redo file structure for java files.
