+++ src/codeu/chat/client/android/app/src/main/AndroidManifest.xml
@@ -10,35 +10,35 @@
         android:label="@string/app_name"
         android:supportsRtl="true"
         android:theme="@style/Theme.Design.NoActionBar">
-        <activity android:name=".LoginActivity">
+        <activity android:name=".view.login.LoginActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".ChatActivity" />
+        <activity android:name=".view.ChatActivity" />
         <activity
-            android:name=".TabsActivity"
+            android:name=".view.TabsActivity"
             android:label="@string/title_activity_users" />
         <activity
-            android:name=".CreateChatActivity"
+            android:name=".view.CreateChatActivity"
             android:label="@string/title_activity_create_chat"
-            android:parentActivityName=".TabsActivity">
+            android:parentActivityName=".view.TabsActivity">
             <meta-data
                 android:name="android.support.PARENT_ACTIVITY"
-                android:value="com.google.codeu.chatme.TabsActivity" />
+                android:value="com.google.codeu.chatme.view.TabsActivity" />
         </activity>
         <activity
-            android:name=".AddUsersActivity"
+            android:name=".view.AddUsersActivity"
             android:label="@string/title_activity_add_users" />
         <activity
-            android:name=".SearchActivity"
+            android:name=".view.SearchActivity"
             android:label="@string/title_activity_search"
-            android:parentActivityName=".ChatActivity">
+            android:parentActivityName=".view.ChatActivity">
             <meta-data
                 android:name="android.support.PARENT_ACTIVITY"
-                android:value="com.google.codeu.chatme.ChatActivity" />
+                android:value="com.google.codeu.chatme.view.ChatActivity" />
         </activity>
     </application>
