+++ src/codeu/chat/client/android/app/src/main/java/com/google/codeu/chatme/view/ChatActivity.java
@@ -1,7 +1,8 @@
-package com.google.codeu.chatme;
+package com.google.codeu.chatme.view;
 
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
+import com.google.codeu.chatme.R;
 
 public class ChatActivity extends AppCompatActivity {
