+++ src/codeu/chat/client/android/app/src/main/java/com/google/codeu/chatme/view/CreateChatActivity.java
@@ -1,7 +1,8 @@
-package com.google.codeu.chatme;
+package com.google.codeu.chatme.view;
 
 import android.os.Bundle;
 import android.app.Activity;
+import com.google.codeu.chatme.R;
 
 public class CreateChatActivity extends Activity {
