+++ src/codeu/chat/client/android/app/src/main/java/com/google/codeu/chatme/view/basicFragments/UserFragment.java
@@ -1,4 +1,4 @@
-package com.google.codeu.chatme;
+package com.google.codeu.chatme.view.basicFragments;
 
 import android.content.Context;
 import android.net.Uri;
@@ -7,6 +7,7 @@ import android.support.v4.app.Fragment;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+import com.google.codeu.chatme.R;
 
 
 /**
