+++ src/codeu/chat/client/android/app/src/main/java/com/google/codeu/chatme/view/login/LoginActivity.java
@@ -1,4 +1,4 @@
-package com.google.codeu.chatme;
+package com.google.codeu.chatme.view.login;
 
 import android.app.ProgressDialog;
 import android.content.Intent;
@@ -11,6 +11,8 @@ import android.widget.Toast;
 
 import com.google.android.gms.tasks.OnCompleteListener;
 import com.google.android.gms.tasks.Task;
+import com.google.codeu.chatme.view.ChatActivity;
+import com.google.codeu.chatme.view.TabsActivity;
 import com.google.firebase.auth.AuthResult;
 import com.google.firebase.auth.FirebaseAuth;
 import com.google.firebase.auth.FirebaseUser;
