+++ src/codeu/chat/client/android/app/src/main/res/layout/activity_chat.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/ChatMe_Red"
-    tools:context="com.google.codeu.chatme.ChatsFragment">
+    tools:context=".view.tabFragments.ChatsFragment">
 
     <!-- TODO: Update blank fragment layout -->
 
@@ -31,7 +31,7 @@
 
     <fragment
         android:id="@+id/fragment4"
-        android:name="com.google.codeu.chatme.MessageFragment"
+        android:name="com.google.codeu.chatme.view.basicFragments.MessageFragment"
         android:layout_width="wrap_content"
         android:layout_height="80dp"
         tools:layout="@layout/fragment_message"
@@ -40,7 +40,7 @@
 
     <fragment
         android:id="@+id/fragment6"
-        android:name="com.google.codeu.chatme.MessageFragment"
+        android:name="com.google.codeu.chatme.view.basicFragments.MessageFragment"
         android:layout_width="wrap_content"
         android:layout_height="80dp"
         android:layout_alignParentStart="true"
@@ -49,7 +49,7 @@
 
     <fragment
         android:id="@+id/fragment5"
-        android:name="com.google.codeu.chatme.MessageFragment"
+        android:name="com.google.codeu.chatme.view.basicFragments.MessageFragment"
         android:layout_width="wrap_content"
         android:layout_height="80dp"
         android:layout_below="@+id/addUserButton"
