+++ src/codeu/chat/client/android/app/src/main/res/layout/activity_create_chat.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/ChatMe_Red"
-    tools:context="com.google.codeu.chatme.ChatsFragment">
+    tools:context=".view.tabFragments.ChatsFragment">
 
     <!-- TODO: Update blank fragment layout -->
