+++ src/codeu/chat/client/android/app/src/main/res/layout/activity_login.xml
@@ -9,7 +9,7 @@
     android:paddingLeft="@dimen/activity_horizontal_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
-    tools:context="com.google.codeu.chatme.LoginActivity">
+    tools:context=".view.login.LoginActivity">
 
     <EditText
         android:id="@+id/password_text"
