+++ src/codeu/chat/client/android/app/src/main/res/layout/fragment_chat.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="80dp"
     android:background="@color/chatMe_Lighter_Red"
-    tools:context="com.google.codeu.chatme.ChatFragment">
+    tools:context=".view.basicFragments.ChatFragment">
 
     <!-- TODO: Update blank fragment layout -->
