+++ src/codeu/chat/client/android/app/src/main/res/layout/fragment_chats.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/ChatMe_Red"
-    tools:context="com.google.codeu.chatme.ChatsFragment">
+    tools:context=".view.tabFragments.ChatsFragment">
 
     <!-- TODO: Update blank fragment layout -->
 
@@ -50,7 +50,7 @@
 
     <fragment
         android:id="@+id/chatFragment2"
-        android:name="com.google.codeu.chatme.ChatFragment"
+        android:name="com.google.codeu.chatme.view.basicFragments.ChatFragment"
         android:layout_width="wrap_content"
         android:layout_height="80dp"
         tools:layout="@layout/fragment_chat"
@@ -59,7 +59,7 @@
 
     <fragment
         android:id="@+id/chatFragment"
-        android:name="com.google.codeu.chatme.ChatFragment"
+        android:name="com.google.codeu.chatme.view.basicFragments.ChatFragment"
         android:layout_width="wrap_content"
         android:layout_height="80dp"
         tools:layout="@layout/fragment_chat"
